requirements.txt
loguru==0.5.3
PyYAML==6.0
setuptools==56.0.0
